pyproject.toml
[project]
name = "xarray-safe-rcm"
requires-python = ">= 3.10"
license = {text = "MIT"}
dependencies = [
    "toolz",
    "numpy",
    "xarray",
    "xarray-datatree",
    "xmlschema",
]
dynamic = ["version"]

[build-system]
requires = ["setuptools>=64.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["safe_rcm"]

[tool.setuptools_scm]
fallback_version = "999"

[tool.isort]
profile = "black"
skip_gitignore = true
float_to_top = true
default_section = "THIRDPARTY"
known_first_party = "safe_rcm"
